修改下chucksize的区间

+++ src/com/bowlong/basic/EOConfig.java
@@ -35,8 +35,8 @@ public class EOConfig {
 	static final public String am_mb = "f:%smb t:%smb u:%smb m:%smb";
 	static final public String am_gb = "f:%sgb t:%sgb u:%sgb m:%sgb";
 	static final public String am_tb = "f:%stb t:%stb u:%stb m:%stb";
-	static final public long kb = 1024; // 存储单位-值
-	static final public long mb = kb * kb;
+	static final public int kb = 1024; // 存储单位-值
+	static final public int mb = kb * kb;
 	static final public long gb = mb * kb;
 	static final public long tb = gb * kb;
 }
